+++ example/lib/main.dart
@@ -19,7 +19,11 @@ import 'package:on_audio_room/on_audio_room.dart';
 void main() async {
   //Init Room.
   await OnAudioRoom().initRoom(RoomType.FAVORITES);
-  runApp(MaterialApp(home: MyApp()));
+  runApp(
+    const MaterialApp(
+      home: MyApp(),
+    ),
+  );
 }
 
 class MyApp extends StatefulWidget {
@@ -31,13 +35,28 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   //Call audio room.
-  OnAudioRoom audioRoom = OnAudioRoom();
+  final OnAudioRoom _audioRoom = OnAudioRoom();
+  final OnAudioQuery _audioQuery = OnAudioQuery();
+
+  @override
+  void initState() {
+    super.initState();
+    _requestPermission();
+  }
+
+  void _requestPermission() async {
+    bool permissionStatus = await _audioQuery.permissionsStatus();
+    if (!permissionStatus) {
+      await _audioQuery.permissionsRequest();
+    }
+    setState(() {});
+  }
 
   @override
   void dispose() {
     //Remember to close room to avoid memory leaks.
     //Choose the better location(page) to add this method.
-    audioRoom.closeRoom();
+    _audioRoom.closeRoom();
     super.dispose();
   }
 
@@ -45,30 +64,31 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("OnAudioRoomExample"),
+        title: const Text("OnAudioRoomExample"),
         actions: [
           IconButton(
-              onPressed: () {
-                Navigator.of(context).push(
-                  MaterialPageRoute(builder: (context) => Information()),
-                );
-              },
-              icon: Icon(Icons.favorite_outline_rounded))
+            onPressed: () {
+              Navigator.of(context).push(
+                MaterialPageRoute(builder: (_) => const Information()),
+              );
+            },
+            icon: const Icon(Icons.favorite_outline_rounded),
+          )
         ],
       ),
       body: FutureBuilder<List<SongModel>>(
         //In this example we will use [on_audio_query] to query songs from device.
-        future: OnAudioQuery().queryAudios(null, null, null, true),
-        builder: (context, item) {
+        future: OnAudioQuery().querySongs(),
+        builder: (_, item) {
           if (item.data != null) {
             List<SongModel> songs = item.data!;
             //Display only 10 because it's only a example.
             return ListView.builder(
-              itemCount: 10,
-              itemBuilder: (context, index) {
+              itemCount: songs.length,
+              itemBuilder: (_, index) {
                 return ListTile(
                   title: Text(songs[index].title),
-                  subtitle: Text(songs[index].artist),
+                  subtitle: Text(songs[index].artist ?? "No artist"),
                   //First method from [on_audio_room], here we'll add the selected song
                   //inside the favorites room.
                   onTap: () async {
@@ -98,21 +118,22 @@ class _MyAppState extends State<MyApp> {
                     //  From your [SongModel] just call [toMap] and then [to*Type*Entity()].
 
                     //This example we use the [on_audio_query] plugin:
-                    var result = audioRoom.addTo(
+                    _audioRoom.addTo(
                       RoomType.FAVORITES, //Specify the room type
                       songs[index].getMap.toFavoritesEntity(),
                     );
-                    print(result);
                   },
                   //Other method, this will clear all the room.
                   onLongPress: () async {
-                    audioRoom.clearRoom(RoomType.FAVORITES);
+                    _audioRoom.clearRoom(RoomType.FAVORITES);
                   },
                 );
               },
             );
           }
-          return Center(child: CircularProgressIndicator());
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
         },
       ),
     );
@@ -134,7 +155,7 @@ class _InformationState extends State<Information> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("OnAudioRoomExample"),
+        title: const Text("OnAudioRoomExample"),
       ),
       body: Center(
         //Add to the future builder the specific type.
@@ -148,9 +169,9 @@ class _InformationState extends State<Information> {
             sortType: null, //Null will use the [key] has sort.
           ),
           builder: (context, item) {
-            if (item.data == null) return CircularProgressIndicator();
+            if (item.data == null) return const CircularProgressIndicator();
 
-            if (item.data!.isEmpty) return Text("No data found");
+            if (item.data!.isEmpty) return const Text("No data found");
 
             List<FavoritesEntity> favorites = item.data!;
             return ListView.builder(
@@ -161,12 +182,12 @@ class _InformationState extends State<Information> {
                   subtitle: Text(favorites[index].dateAdded.toString()),
                   onTap: () async {
                     //Most the method will return a bool to indicate if method works.
-                    var result = await audioRoom.deleteFrom(
-                        RoomType.FAVORITES, favorites[index].key);
+                    await audioRoom.deleteFrom(
+                      RoomType.FAVORITES,
+                      favorites[index].key,
+                    );
                     //Call setState to see the result,
-                    setState(() {
-                      print(result);
-                    });
+                    setState(() {});
                   },
                 );
               },
@@ -175,25 +196,5 @@ class _InformationState extends State<Information> {
         ),
       ),
     );
-  }
-}
-
-//This method is temporary, soon will be added to [on_audio_query].
-//TODO: Add this extension to [on_audio_query] and [toString()].
-extension OnAudioFormatter on SongModel {
-  Map get getMap {
-    Map tempMap = {};
-    tempMap["_data"] = this.data;
-    tempMap["_display_name"] = this.displayName;
-    tempMap["_id"] = this.id;
-    tempMap["album"] = this.album;
-    tempMap["album_id"] = this.albumId;
-    tempMap["artist"] = this.artist;
-    tempMap["artist_id"] = this.artistId;
-    tempMap["date_added"] = this.dateAdded;
-    tempMap["duration"] = this.duration;
-    tempMap["title"] = this.title;
-    tempMap["artwork"] = this.artwork;
-    return tempMap;
   }
 }