+++ lib/details/extensions/room_helper_extension.dart
@@ -1,7 +1,7 @@
 import 'package:on_audio_room/on_audio_room.dart';
 
 /// A [on_audio_room] extension to help the package.
-extension OnSongsHelper on List<SongEntity> {
+extension OnSongsListHelper on List<SongEntity> {
   /// Used to get all [ids] from a list of [SongEntity].
   List<int> get getAllIds {
     List<int> tempKeyList = [];
@@ -10,10 +10,42 @@ extension OnSongsHelper on List<SongEntity> {
     });
     return tempKeyList;
   }
+
+  /// A helper to get a list of [SongEntity] as a [Map].
+  List<Map> get getListOfMap {
+    List<Map> tempList = [];
+
+    this.forEach((entity) {
+      tempList.add(entity.getMap);
+    });
+
+    return tempList;
+  }
 }
 
 /// A [on_audio_room] extension to help the package.
-extension OnFavoritesHelper on List<FavoritesEntity> {
+extension OnSongsHelper on SongEntity {
+  /// A helper to get the [SongEntity] as a [Map].
+  Map get getMap {
+    Map tempKeyMap = {
+      "_data": this.lastData,
+      "_display_name": this.displayName,
+      "_id": this.id,
+      "album": this.album,
+      "album_id": this.albumId,
+      "artist": this.artist,
+      "artist_id": this.artistId,
+      "date_added": this.dateAdded,
+      "duration": this.duration,
+      "title": this.title,
+      "artwork": this.artworkAsBytes,
+    };
+    return tempKeyMap;
+  }
+}
+
+/// A [on_audio_room] extension to help the package.
+extension OnFavoritesListHelper on List<FavoritesEntity> {
   /// Used to get all [keys] from a list of [FavoritesEntity].
   List<int> get getAllKeys {
     List<int> tempKeyList = [];
@@ -32,10 +64,42 @@ extension OnFavoritesHelper on List<FavoritesEntity> {
     });
     return tempMap;
   }
+
+  /// A helper to get a list of [FavoritesEntity] as a [Map].
+  List<Map> get getListOfMap {
+    List<Map> tempList = [];
+
+    this.forEach((entity) {
+      tempList.add(entity.getMap);
+    });
+
+    return tempList;
+  }
 }
 
 /// A [on_audio_room] extension to help the package.
-extension OnLastPlayedHelper on List<LastPlayedEntity> {
+extension OnFavoritesHelper on FavoritesEntity {
+  /// A helper to get the [FavoritesEntity] as a [Map].
+  Map get getMap {
+    Map tempKeyMap = {
+      "_data": this.lastData,
+      "_display_name": this.displayName,
+      "_id": this.id,
+      "album": this.album,
+      "album_id": this.albumId,
+      "artist": this.artist,
+      "artist_id": this.artistId,
+      "date_added": this.dateAdded,
+      "duration": this.duration,
+      "title": this.title,
+      "artwork": this.artworkAsBytes,
+    };
+    return tempKeyMap;
+  }
+}
+
+/// A [on_audio_room] extension to help the package.
+extension OnLastPlayedListHelper on List<LastPlayedEntity> {
   /// Used to get all [keys] from a list of [LastPlayedEntity].
   List<int> get getAllKeys {
     List<int> tempKeyList = [];
@@ -54,10 +118,42 @@ extension OnLastPlayedHelper on List<LastPlayedEntity> {
     });
     return tempMap;
   }
+
+  /// A helper to get a list of [LastPlayedEntity] as a [Map].
+  List<Map> get getListOfMap {
+    List<Map> tempList = [];
+
+    this.forEach((entity) {
+      tempList.add(entity.getMap);
+    });
+
+    return tempList;
+  }
 }
 
 /// A [on_audio_room] extension to help the package.
-extension OnMostPlayedHelper on List<MostPlayedEntity> {
+extension OnLastPlayedHelper on LastPlayedEntity {
+  /// A helper to get the [LastPlayedEntity] as a [Map].
+  Map get getMap {
+    Map tempKeyMap = {
+      "_data": this.lastData,
+      "_display_name": this.displayName,
+      "_id": this.id,
+      "album": this.album,
+      "album_id": this.albumId,
+      "artist": this.artist,
+      "artist_id": this.artistId,
+      "date_added": this.dateAdded,
+      "duration": this.duration,
+      "title": this.title,
+      "artwork": this.artworkAsBytes,
+    };
+    return tempKeyMap;
+  }
+}
+
+/// A [on_audio_room] extension to help the package.
+extension OnMostPlayedListHelper on List<MostPlayedEntity> {
   /// Used to get all [keys] from a list of [MostPlayedEntity].
   List<int> get getAllKeys {
     List<int> tempKeyList = [];
@@ -75,5 +171,37 @@ extension OnMostPlayedHelper on List<MostPlayedEntity> {
       tempMap[entity.key] = entity;
     });
     return tempMap;
+  }
+
+  /// A helper to get a list of [MostPlayedEntity] as a [Map].
+  List<Map> get getListOfMap {
+    List<Map> tempList = [];
+
+    this.forEach((entity) {
+      tempList.add(entity.getMap);
+    });
+
+    return tempList;
+  }
+}
+
+/// A [on_audio_room] extension to help the package.
+extension OnMostPlayedHelper on MostPlayedEntity {
+  /// A helper to get the [MostPlayedEntity] as a [Map].
+  Map get getMap {
+    Map tempKeyMap = {
+      "_data": this.lastData,
+      "_display_name": this.displayName,
+      "_id": this.id,
+      "album": this.album,
+      "album_id": this.albumId,
+      "artist": this.artist,
+      "artist_id": this.artistId,
+      "date_added": this.dateAdded,
+      "duration": this.duration,
+      "title": this.title,
+      "artwork": this.artworkAsBytes,
+    };
+    return tempKeyMap;
   }
 }