+++ lib/details/rooms/favorites/favorites_entity.g.dart
@@ -29,13 +29,14 @@ class FavoritesEntityAdapter extends TypeAdapter<FavoritesEntity> {
       ..dateAdded = fields[7] as int?
       ..duration = fields[8] as int?
       ..title = fields[9] as String
-      ..artwork = fields[10] as String?;
+      ..artwork = fields[10] as String?
+      ..artworkAsBytes = fields[15] as Uint8List?;
   }
 
   @override
   void write(BinaryWriter writer, FavoritesEntity obj) {
     writer
-      ..writeByte(12)
+      ..writeByte(13)
       ..writeByte(11)
       ..write(obj.key)
       ..writeByte(0)
@@ -59,7 +60,9 @@ class FavoritesEntityAdapter extends TypeAdapter<FavoritesEntity> {
       ..writeByte(9)
       ..write(obj.title)
       ..writeByte(10)
-      ..write(obj.artwork);
+      ..write(obj.artwork)
+      ..writeByte(15)
+      ..write(obj.artworkAsBytes);
   }
 
   @override